Fix broken tests caused by test helper/API incompatibility
The addon we use for our feature flag implementation
(ember-feature-flags) ships with a test helper (withFeature) that we use
to toggle a feature for a given test.

Unfortunately, it isn't compatible with how we were setting feature
flags via the API. We were using the addon's setup method, which would
reset the state of the feature flags, causing our feature flags set in
the tests to be unset and our tests to break.

Given that `setup` is destructive, the fix is to simply enable/disable
based on the API response.  This would still be destructive if there
were overlap between what we return in the mocked out API response and
a feature toggled with the test helper, but we can deal with that later
if/when it arises.